requirements.txt
absl-py==0.4.1
annoy==1.13.0
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.14.2
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.1
pyemd==0.5.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
